+++ Ranger/Services/SharkScope/Metadata/Metadata.swift
@@ -1,25 +1,26 @@
 //
-//  SharkScopeResponse.swift
+//  PlayerSummary.swift
 //  Ranger
 //
-//  Created by Geri Borbás on 2019. 12. 12..
+//  Created by Geri Borbás on 2019. 12. 16..
 //  Copyright © 2019. Geri Borbás. All rights reserved.
 //
 
 import Foundation
 
 
-struct Metadata: Decodable
+struct Metadata: RootResponse
+{ let Response: MetadataResponse }
+
+
+struct MetadataResponse: Response
 {
-    
-    
-    let Response: Response
-    
-    
-    struct Response: Decodable
-    {
-        
-        
-        let metadataHash: String
-    }
+
+
+    let metadataHash: String
+    let timestamp: StringFor<Date>
+    let success: StringFor<Bool>
+
+    // let MetadataResponse: Any
+    let UserInfo: UserInfo
 }